added xfonts-base and tightvncserver deb files
As installing a VNC server is a very popular and common Raspberry Pi setup task, it is also a task that many people would like to perform offline. In order for this to be possible, the tightvncserver deb file must be distributed with the "install vnc server" block. Luckily this package only depends on one other that is not installed by default: xfonts-base, so this one has also been added to the block.