removed reboot and added network
The “reboot” option for blocks has been removed.
A new option - “network” - has been added. This allows the creator of
the block to specify whether the block requires a network connection to
work. Note that the block will always be run, even if there is no
network connection - the “network” parameter is just for PiBakery to
know if it should spend time waiting for a network connection.