updates otgether to be Stretch compatible
Stretch broke the static IP portion of this script. It appears that dhcpcd.conf doesn't accept static IPs in the 169.254 range - when attempting to set this it simply didn't set an ipv4 address for the interfact (usb0 in the case of this script). See https://www.raspberrypi.org/forums/viewtopic.php?t=199116 for more info.

The new method of setting a static IP involves adding a dhcpcd exit hook to set the IP of the usb0 interface. A bit hacky, but it works on Stretch.